+++ frontend/src/pages/PackSimulator/PackOpen/PackOpen.styles.ts
