+++ frontend/src/pages/PackSimulator/PackSelect/PackSelect.styles.ts
