+++ frontend/src/pages/PackSimulator/PackSelect/components/Pack/Pack.styles.ts
