+++ frontend/src/pages/PackSimulator/PackSelect/components/Pack/Pack.tsx
