+++ frontend/src/pages/PackSimulator/PackSimulator.style.ts
