+++ frontend/src/pages/PackSimulator/StatisticsInfo/StatisticContent/StatisticContent.style.ts
