+++ frontend/src/pages/PackSimulator/StatisticsInfo/StatisticContent/StatisticItem.style.ts
