+++ frontend/src/pages/PackSimulator/StatisticsInfo/StatisticContent/StatisticItem.tsx
