+++ frontend/src/pages/PackSimulator/StatisticsInfo/StatisticsInfo.styles.ts
