+++ frontend/src/pages/PackSimulator/importImages.ts
